refactor(boot): extract long mode transition to entry.S and share bootlib
- Extract enter_long_mode() inline asm from bootload.c into boot/bios/entry.S
  as a standalone assembly routine for better readability and maintainability
- Create boot/bootlib.h with shared memcpy/memset/memcmp/strcpy for both
  BIOS (32-bit) and UEFI (64-bit) bootloaders, eliminating duplication
- Fix AHCI driver: use dynamic PCI BAR discovery instead of hardcoded base,
  fix s_ahci_base type (uint32_t -> uintptr_t) and shadowed local variable
- Improve IDE detection: validate DRQ/ERR status and reject zero-size devices
- Add AHCI controller to QEMU configs (qemu.cfg, qemu-uefi.cfg) for SATA
  disk testing alongside IDE boot disk
- Fix types.h: guard C++ templates with #ifdef __cplusplus, add C fallbacks
- Merge adjacent UEFI memory map entries to reduce mmap table size
- Makefile: split bootloader into separate C + asm object files, add
  fat32_test.img dependency to qemu/qemu-uefi targets

+++ kernel/drivers/ahci.h
@@ -150,5 +150,5 @@ private:
     static int s_ahci_devices_count;
 
     static AhciPortConfig s_ahci_port_configs[ahci::MAX_DEVICES];
-    static uint32_t s_ahci_base;  // AHCI controller MMIO base
+    static uintptr_t s_ahci_base;  // AHCI controller MMIO virtual base
 };
